+++ device_preview/lib/src/plugins/shared_preferences_explorer.dart
@@ -76,6 +76,7 @@ class _AllPreferencesViewState extends State<_AllPreferencesView> {
           ),
         ),
       );
+      widget.onKeySelected(null);
     }
   }